+++ pkgs/jnigen/example/kotlin_plugin/lib/kotlin_bindings.dart
@@ -102,20 +102,23 @@ class Example extends jni$_.JObject {
             reference.pointer,
             _id_thinkBeforeAnswering as jni$_.JMethodIDPtr,
             _$continuation.pointer)
-        .object<jni$_.JObject>(const jni$_.JObjectType())
-        .reference;
+        .object<jni$_.JObject>(const jni$_.JObjectType());
     _$continuation.release();
-    final jni$_.JReference $o;
-    if (jni$_.Jni.env.IsInstanceOf(
-      $r.pointer,
-      jni$_.coroutineSingletonsClass.reference.pointer,
-    )) {
-      $o = jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
+    final jni$_.JObject $o;
+    if ($r.isInstanceOf(jni$_.coroutineSingletonsClass)) {
       $r.release();
+      $o = jni$_.JObject.fromReference(
+          jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first)));
+      if ($o.isInstanceOf(jni$_.result$FailureClass)) {
+        final $e =
+            jni$_.failureExceptionField.get($o, const jni$_.JObjectType());
+        $o.release();
+        jni$_.Jni.throwException($e.reference.toPointer());
+      }
     } else {
       $o = $r;
     }
-    return const jni$_.JStringType().fromReference($o);
+    return $o.as(const jni$_.JStringType(), releaseOriginal: true);
   }
 }