+++ pkgs/jnigen/test/kotlin_test/kotlin/src/main/kotlin/com/github/dart_lang/jnigen/SuspendFun.kt
@@ -18,6 +18,7 @@ public class SuspendFun {
     }
 
     suspend fun fail(): String {
+        delay(100L)
         throw Error()
     }